refactor(mcp): split authnchallenge.go per handler
bflad PR #2751 review: 1651-line file is hard to navigate. Split by
handler with shared core types/helpers staying in authnchallenge.go:

  authnchallenge.go              — AuthnChallengeState, UserSessionGrant,
                                   WriteAuthenticateChallenge,
                                   validateUserSessionToken,
                                   requireUserSessionIssuer,
                                   extractClientCredentials, sha256Hex,
                                   generateOpaqueToken
  authnchallenge_well_known.go    — RFC 9728 + RFC 8414 metadata.
  authnchallenge_register.go      — RFC 7591 DCR.
  authnchallenge_authorize.go     — RFC 6749 §4.1.1 authorization.
  authnchallenge_idp_callback.go  — Speakeasy IDP callback.
  authnchallenge_consent.go       — consent UI + POST.
  authnchallenge_token.go         — RFC 6749 §4.1.3 / §6 token.
  authnchallenge_revoke.go        — RFC 7009 token revocation.

Pure code move + import-block split per file. No behaviour changes.
